vim-6.4a release

+++ src/mbyte.c
@@ -3044,7 +3044,7 @@ iconv_string(vcp, str, slen, unconvlenp)
 	    else
 	    {
 		l = (*mb_ptr2len_check)((char_u *)from);
-		if (l > fromlen)
+		if ((size_t)l > fromlen)
 		    l = fromlen;
 	    }
 	    from += l;
